Move config information out of setup file
