+++ loko/__init__.py
@@ -1,21 +1,9 @@
 import os
 import logging
 from logging.handlers import RotatingFileHandler
-import configparser
-from sqlalchemy.orm import sessionmaker
-
-
-CONFIG_FILENAME = 'loko_config.cfg'
-
-def get_configs(profile = 'DEFAULT'):
-	
-	curdir = os.path.dirname(__file__)	#The Path to this __init__ file.
-	curdir = os.path.abspath(curdir)
-	config = configparser.ConfigParser()
-	config.read(os.path.join(curdir,'..',CONFIG_FILENAME))
-	
-	return config[profile]
-
+from loko.conf import get_configs
+from loko import models
+from loko.models.models import Base
 
 def get_logger(loggerName):
 	log = logging.getLogger(loggerName)
@@ -38,3 +26,40 @@ def get_logger(loggerName):
 	log.setLevel(logging.DEBUG)
 	return log
 
+#TODO: Move to conf.py
+class LokoConfig:
+	def __init__(self, engine, logger):
+		self.logger = logger
+		# self.newconfig(engine)
+
+	def create_tables(self, engine):
+		print("Getting the tables")
+		tables = models.get_tables_metadata()
+		for table in tables:
+			print("Creating table *{}*".format(table.name))
+			self.logger.info("Creating table *{}*".format(table.name))
+			#create_all will not recreate existing tables
+			Base.metadata.create_all(engine, tables=[table], checkfirst=True)
+			self.logger.info("Table {} created successfully".format(table.name))
+
+	def drop_tables(self, engine):
+		tables = models.get_tables_metadata()
+		tables.reverse()
+		self.logger.warning("Deleting all tables")
+		#Do dangerous stuff here and log as each table is deleted
+		for table in tables:
+			self.logger.warning("Deleting table *{}*".format(table.name))
+			#TODO: Move use of Base away from here to somewhere in loko.models
+			Base.metadata.drop_all(engine, tables=[table])
+			self.logger.info("Table *{}* deleted successfully".format(table.name))
+
+		self.logger.info("Tables deleted successfully!")
+
+	def newconfig(self, engine, drop_tables=False):	#For a new install
+		'''
+		if drop_tables, drop all tables and create afresh, otherwise
+		#only create new tables
+		'''
+		if drop_tables:
+			self.drop_tables(engine)
+		self.create_tables(engine)
